Refactor Makefile and add parsing_here_doc.c
- Refactor Makefile to include parsing_here_doc.c
- Add parsing_here_doc.c file to handle here documents in commands
